1.1
revert

+++ src/networkmessage.cpp
@@ -1,6 +1,4 @@
 /**
- * @file networkmessage.cpp
- * 
  * The Forgotten Server - a free and open-source MMORPG server emulator
  * Copyright (C) 2019 Mark Samman <mark.samman@gmail.com>
  *
@@ -28,7 +26,7 @@
 
 int32_t NetworkMessage::decodeHeader()
 {
-	int32_t newSize = buffer[0] | buffer[1] << 8;
+	int32_t newSize = static_cast<int32_t>(buffer[0] | buffer[1] << 8);
 	info.length = newSize;
 	return info.length;
 }
@@ -75,7 +73,7 @@ void NetworkMessage::addString(const std::string& value)
 void NetworkMessage::addDouble(double value, uint8_t precision/* = 2*/)
 {
 	addByte(precision);
-	add<uint32_t>((value * std::pow(static_cast<float>(10), precision)) + std::numeric_limits<int32_t>::max());
+	add<uint32_t>(static_cast<uint32_t>((value * std::pow(static_cast<float>(10), precision)) + std::numeric_limits<int32_t>::max()));
 }
 
 void NetworkMessage::addBytes(const char* bytes, size_t size)
